server sends mssg client missing threads

+++ server.py
@@ -5,19 +5,44 @@ import threading
 
 
 def processEntry(data, addr):
-    d = data.replace('\\', ' ')
+    user_exists = 0
+    user_ok = 0
+    d = data.replace('\\', ' ')  # format string
     d_json = json.loads(d)
-    d_json['id'] = addr[0]
-    print(d_json)
-    #solve_action(d_json)
-    with open('example.json', 'a+') as f:
-        f.seek(0, 2)
-        f.truncate()
-        f.seek(-2, 2)
-        f.truncate()
-        f.write(' , ')
-        json.dump(d_json, f)
-        f.write(']}')
+    d_json['id'] = addr      # change id to ip address
+    # verify that usernames are unique
+    data = json.load(open('example.json'))
+    if d_json['action'] == 'e':  # if a new user is registering
+        print("new user registering....")
+        for d in data['messages']:
+            if (d['user'] == d_json['user']) and (d_json['action'] == 'e'):
+                user_exists = 1
+            elif (d['user'] != d_json['user']) and (d_json['action'] == 'e'):
+                user_ok = 1
+        if user_exists == 1:
+            s.sendto("Username already exist", d_json['id'])
+        elif user_ok == 1:
+            s.sendto("Username valid", d_json['id'])
+            with open('example.json', 'a+') as f:  # adds user to json
+                f.seek(0, 2)
+                f.truncate()
+                f.seek(-2, 2)
+                f.truncate()
+                f.write(' , ')
+                json.dump(d_json, f)
+                f.write(']}')
+    #  TODO update each users must recent action
+    else:
+        with open('example.json', 'a+') as f:
+            f.seek(0, 2)
+            f.truncate()
+            f.seek(-2, 2)
+            f.truncate()
+            f.write(' , ')
+            json.dump(d_json, f)
+            f.write(']}')
+        solve_action(d_json)
+
 
 
 def give_users():
@@ -39,8 +64,18 @@ def broadcast_message(d_json):
 
 
 def private_message(d_json):
-    print("*PRIVATE MESSAGE")
-    s.sendto(d_json['text'], d_json['id'])
+    error_mssg = "User doesn't exist"
+    text = d_json['text']
+    array_text = text.split(',')
+    private_message = array_text[0]
+    recipient_pm = array_text[1]
+    data = json.load(open('example.json'))
+    for d in data['messages']:
+        if recipient_pm == d['user']:
+            print("Private message sent: ", private_message, tuple(d['id']))
+            s.sendto(private_message, tuple(d['id']))
+        else:
+            s.sendto(error_mssg, d_json['id'])
 
 
 def exit(d_json):
@@ -59,7 +94,8 @@ def solve_action(d_json):
         private_message(d_json)
     elif action == 'd':
         exit(d_json)
-
+    elif action == 'e':
+        print("new user saved")
 
 HOST = ''   # Symbolic name meaning all available interfaces
 PORT = 8888 # Arbitrary non-privileged port
@@ -93,7 +129,6 @@ while 1:
 
     reply = 'OK...' + data
     processEntry(data, addr)
-
     s.sendto(reply, addr)
     print 'Message[' + addr[0] + ':' + str(addr[1]) + '] - ' + data.strip()
